In the BSD standard I/O library, there's a flag in a FILE structure that's set whenever we encounter an EOF; if that flag is set, all subsequent reads return an EOF indication.  I.e., end-of-file is sticky.
This means that the stuff to continue reading a capture file, if we're
updating the display as the capture progresses, doesn't work - it gets
stuck at the point where the first read finished.

To clear that flag, we must do an "fseek()"; we do one that doesn't move
the seek pointer.

When updating the display as a capture progresses, do
"init_col_widths()" only when we first open the capture file; there's no
need to do it every time we read from the file - the column widths never
get smaller, they can only get bigger or stay the same.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@370 f5534014-38df-0310-8fa8-9805f1628bb7