Add a "File/Print" menu item, which prints *all* the packets in the capture to a file or printer.  This should eventually get the ability to print either all the packets or only the packets selected by the display filter, and possibly also the ability to print only packets M through N.
Get rid of "cur" member of "capture_file" structure; nobody used it.

There's no need to pass a pointer to a "dialog_button" variable to
"simple_dialog()" for the error boxes displayed if a file copy or move
fails; that dialog box is just a message box and has only an "OK"
button.

Put the declaration of "prefs" into "prefs.h".


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@378 f5534014-38df-0310-8fa8-9805f1628bb7