Olivier Abad's patch to add dissectors for LAP-B and X.25, and wiretap support for RADCOM Ltd.'s WAN/LAN analyzers (see
	http://www.radcom-inc.com/

).  Note: a

Make "S" a mnemonic for "Summary" in the "Tools" menu.

Move the routine, used for the "Tools/Summary" display, that turns a
wiretap file type into a descriptive string for it into the wiretap
library itself, expand on some of its descriptions, and add an entry for
files from a RADCOM analyzer.

Have "Tools/Summary" display the snapshot length for the capture.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@417 f5534014-38df-0310-8fa8-9805f1628bb7