Add the ability to specify a filter to be used when reading the file to the "Open File" dialog box (the "Open File" dialog box equivalent of the "-R" flag).  Have "load_cap_file()" take the filter expression as an argument, and make the global "rfilter" into a member of a "capture_file" structure.
When reading a temporary capture file after a live capture, don't apply
any filter.

Move the code that pops up error boxes on file opens when reading a
capture file back to "load_cap_file()"; it also pops up error boxes if
the filter expression can't be parsed.

Don't enable "File/Save" or "File/Save As..." if an attempt to read a
capture file fails - if there was already an open capture file, it was
closed by "load_cap_file()", so we no longer have an open file to save.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@460 f5534014-38df-0310-8fa8-9805f1628bb7