Pick up a change from the OpenBSD ports collection, to use "mkstemp()" to create the temporary file to which the "Follow TCP Stream" code writes the contents of the TCP stream; this was presumably done to close a potential security hole opened by the possibility of somebody replacing the file referred to by the name generated by "tmpnam()" before we open it.
(We may have to provide our own version of "mkstemp()" if there are any
platforms on which we run that lack it.  Note that Solaris 2.5.1 has it,
even though it's not documented and isn't declare in its "stdlib.h".)

Have the "Follow TCP Stream" code put up a dialog box, rather than
spitting a message out to the standard error, if the attempt to create
the temporary file fails.

Check also whether the attempt to open the "Follow TCP Stream" temporary
file for reading fails, and put up a dialog box for that as well.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@502 f5534014-38df-0310-8fa8-9805f1628bb7