"read_filters()" is actually called late enough in the startup process that it can pop up a message box; have it do so.
Make the "Can't open file" message boxes in "colors.c" include the
"errno" error message in the message they put up.

Don't complain about being unable to open the color filter file if it
doesn't exist (perhaps the user just never made one).

Make the message for a failure to open the preferences file resemble
that for a failure to open a color filter file.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@575 f5534014-38df-0310-8fa8-9805f1628bb7