PPP options in LCP, IPCP, etc. are like IP and TCP options - one octet of option code, one octet of length (which includes the two option code and length bytes), followed by 0 or more octets of option data, with some options being fixed-length and some being variable-length.  Put some stuff from the PPP control protocol option parsing code into the IP-and-TCP option parsing code, and use the latter instead of the former.
(That code might also be usable for CDP as well, with some stuff added
to it.)

Shuffle the arguments to "dissect_ip_tcp_options()" to resemble those of
various other dissectors (i.e., with the "proto_tree *" at the end).

Add in code to dissect a pile of PPP options documented in various RFCs.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@601 f5534014-38df-0310-8fa8-9805f1628bb7