A child process for a "-S" or "-F" capture should *always* exit when the capture is done; make it do so, and don't bother passing it a "-Q" flag to tell it to do so.
"capture()" is called in two places; in one place, it's in a child
process, and it shouldn't read in the capture file.  Move the reading of
the capture file out of "capture()" itself to the place where we
*should* read in the capture file after it returns.  Also, have it
return an indication of whether it succeeded or failed, so we know
whether we should read in the capture file.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@747 f5534014-38df-0310-8fa8-9805f1628bb7