Have "dfilter_compile()" return 0 on success and 1 on failure, and return the pointer to the compiled filter through a pointer argument.
Have it check whether the filter is a null filter and, if so, free up
the filter and supply a filter pointer, rather than obliging its callers
to check whether the filter actually has any code.  (Well, they may want
to check if the filter is null, so that they don't save a pointer to the
filter text, e.g. so that the display filter displays as "none" rather
than as a blank string in the summary box.)

In the process, fix the check in "gtk/file_dlg.c" that tests whether the
read filter compiled successfully.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@812 f5534014-38df-0310-8fa8-9805f1628bb7