New proto_tree header_field_info stuff. Header_field_infos now contain the base for numbers to be displayed in, bitmasks for bitfields, and blurbs (which are one or two sentences describing the field).
proto_tree_add*() routines now automatically handle bitfields. You tell
it which header field you are adding, and just pass it the value of the
entire field, and the proto_tree routines will do the masking and shifting
for you.

This means that bitfields are more naturally filtered via dfilter now.

Added Phil Techau's support for signed integers in dfilters/proto_tree.

Added the beginning of the SNA dissector. It's not complete, but I'm
committing it now because it has example after example of how to use
bitfields with the new header_field_info struct and proto_tree routines.
It was the impetus to change how header_field_info works.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@815 f5534014-38df-0310-8fa8-9805f1628bb7