Add a "Find Frame" menu item under "Display"; it lets you use a display filter to search forward or backward in the list of displayed frames for a matching frame.
When filtering the display, readjust the display to show the "current"
frame if it passed the display filter.  When a file is read in, the
first frame becomes the "current" frame; when a frame is selected, it
becomes the "current" frame, and remains so *even if you unselect it*,
until another frame is selected.

Select the first frame when a file is read in.

Disable most of the "Display" and "Tools" menu items if there's no
current capture file, and enable the relevant ones if there is.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@983 f5534014-38df-0310-8fa8-9805f1628bb7