It's now called "editcap", as it can read any file format Wiretap can read, and write any format it can write; change the error messages.
Make the "-F" flag take a symbolic capture file type; use
"wtap_short_string_to_file_type()" to translate it to a WTAP_FILE_
value.

List, in the usage message, the capture file types we can write, and the
symbolic types you use on the command line to specify them.

Give it an RCS ID.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@1210 f5534014-38df-0310-8fa8-9805f1628bb7