More patches for packet-smb.c ...
Now should be decoding the names of lots more LanMan API request. These
were culled from Samba.  Would be good to go through and give names to the
fields as well.

Will soon decode the response structures returned and then will look at
ways to specify that built-in routines should be called to decode an element.

I also need some captures with UNICODE in them. Anyone got any? Someone
sent in a patch for UNICODE handling, but I did not realize what it was and
now the code has diverged so far it is hard to apply the patch ...

Send captures to rsharpe@ns.aus.com./


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@1334 f5534014-38df-0310-8fa8-9805f1628bb7