Add a new page to the Preferences notebook: a GUI page. The sole option right now is the placement of the vertical scrollbars in the 3 panes. (it's one decision; you can't have the placement of the vertical scrollbar in the packet list pane different than the placement in the protocol tree pane, for example).
I did this because I find it convenient to have the vertical scrollbars
on the *left* side of the text. My mouse cursor is usually expanding and
collapsing the protocol tree widgets, and once the protocol tree changes
size, I usually have to scroll. I'd rather move my mouse cursor just a few
pixels over to find the vertical scrollbar.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@1351 f5534014-38df-0310-8fa8-9805f1628bb7