Put "Ethereal" in boldface everywhere it appears.
Reformat some (source) paragraphs, for the benefit of those with editors
that don't wrap lines in the display.

Delete some extra "=back" directives.

Fix the description of the "Preferences" dialog (it lets you control
various preferences, not just print preferences; it's the "Print" tab
that lets you control print preferences).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@1352 f5534014-38df-0310-8fa8-9805f1628bb7