Don't bother checking whether the filter string is empty - just use it even if it's empty, rather than using an empty string instead of it if it's empty.
Also, "gtk_entry_get_text()" appears, from the documentation, always to
return a non-null pointer (and it returned a non-null pointer when I
started a capture with an empty filter string); don't supply a null
string if it's null, just do a "g_assert()" to make sure it's non-null.

Put in a comment explaining why we aren't trying to be clever and
detecting an empty filter string and setting "cf.filter" to NULL if it's
empty.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@1423 f5534014-38df-0310-8fa8-9805f1628bb7