Fix some errors discovered by making GCC do format string/argument cross-checking, and by replacing "proto_tree_add_item_format()" by multiple routines to add items of various types.
Make the arguments of "proto_tree_add_bytes_format()" and
"proto_tree_add_string_format()" that specify the bytes or the string be
"const" pointers, so that one can pass a "const" pointer without
complaints from the compiler.

Squelch a (bogus, but the compiler isn't in a position to know that)
complaint about an uninitialized variable.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@1716 f5534014-38df-0310-8fa8-9805f1628bb7