In a Network Monitor capture file, get the starting offsets of frames from the frame table - Network Monitor 2.x, at least, doesn't always write frame N+1 right after frame N.
To do that, we need to mallocate a big array to hold the frame table,
and free it when we close the capture file; this requires that we have
capture-file-type-specific close routines as well as
capture-file-type-specific read routines - we let it the pointer to that
routine be null if it's not needed.  Given that, we might as well get
rid of the switch statement in "wtap_close()", in favor of using
capture-file-type-specific close routines, as per the comment before
that switch statement.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@1740 f5534014-38df-0310-8fa8-9805f1628bb7