Set the COL_PROTOCOL column to "ISIS" in the ISIS dissector, rather than in its caller.
Set the COL_PROTOCOL column to "COTP" in the COTP dissector, rather than
in the dissectors it calls, so that it's done for all frames we hand to
that dissector.

Set the COL_INFO field for COTP packets we don't dissect because the
length indicator is zero, or the frame is too short, or the PDU type is
bogus.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@1842 f5534014-38df-0310-8fa8-9805f1628bb7