If an NPDU contains multiple concatenated TPDUs, dissect all of them.
Fix up some spacing problems in the INFO field of CLNP packets.

Dissect only the payload of DT and MD NPDUs as COTP; the payload of
other packets isn't to be so dissected (ER NPDUs are like ICMP error
packets, in that they contain the network-layer header and zero or more
bytes of the payload of the offending packet, for example).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@1860 f5534014-38df-0310-8fa8-9805f1628bb7