Use "BYTES_ARE_IN_FRAME()" rather than explicitly checking an offset and packet length.
Use "IS_DATA_IN_FRAME()" rather than checking the value of
"END_OF_FRAME" when checking whether there's any data left in the
packet.

Before putting the initial login sequence, or any part of it, into the
tree, make sure it's actually in the packet.

When looking for the end of a '\0'-terminated string, don't run past the
end of the captured data in the frame.

Before putting the terminal information into the tree, make sure it's
actually in the packet.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@2215 f5534014-38df-0310-8fa8-9805f1628bb7