In TCP Follow window, allow the optional showing of:
	Entire Conversation
	Client -> Server packets
	Server -> Client packets

Have "Save As" button work as a "Print to File" button; it asks for
a filename and uses the same routine that "Print" uses to save the file.
What you see in the window is what you get in the file. So, you can get
any of the above conversations/soliloquies combined with:

	ASCII
	EBCDIC
	Hex Dump


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@2232 f5534014-38df-0310-8fa8-9805f1628bb7