Give them RCS IDs.
Fix a typo in the LGPL in the initial comment.

Move the includes of <stdlib.h> and <ctype.h> out of "snprintf-imp.h"
into "snprintf.c", and put them *before* the include of "snprintf.h", so
that "size_t" is defined before "snprintf.h" is included ("snprintf.h"
uses "size_t").


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@2257 f5534014-38df-0310-8fa8-9805f1628bb7