This is an SMTP dissector, not a BXXP dissector; fix the comment at the beginning.
Pass "pinfo->fd", not "fd", to "p_get_proto_data()", so that it'll
continue to work even when tvbuffified.

Use "strchr()", not "index()" - "strchr()" is in the ANSI C standard,
and may be in some systems that don't have "index()", whereas those
systems that had "index()" but not "strchr()" got with the ANSI C
program a while ago.

Use "old_dissector_add()" and "old_dissector_delete()" to register and
unregister the SMTP dissector, as it's not yet been tvbuffified.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@2303 f5534014-38df-0310-8fa8-9805f1628bb7