Use "memset()" to zero byte arrays, and "memcpy" to copy into them.
When copying IPv4 or IPv6 addresses, copy the minimum of the number of
bytes in the source and the destination, so that if the prefix length is
bogus, we don't overwrite the buffer.

"tvb_get_ptr()" returns a "guint8 *"; give the variable to which its
value is assigned that type.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@2307 f5534014-38df-0310-8fa8-9805f1628bb7