Change from Andreas Sikkema to change the "old" and "new" members of the "dissector" union in the "conversation_t" structure to "old_d" and "new_d", to avoid using a C++ reserved word.
Add "old_conversation_set_dissector()" and
"conversation_set_dissector()" routines to set the dissector for a
conversation, to hide the details of how that's done (e.g., details such
as whether there's a union at all - eventually, when all dissectors have
been tvbuffified, there won't be a need for the union - and what the
names of the union members are, and so on).  Convert all dissectors to
use those routines (they had to be changed anyway, due to the name
change).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@2324 f5534014-38df-0310-8fa8-9805f1628bb7