Add a help context identifier for the status bar.
This allows the display of a short field description (and
field abbreviation) in the status bar when a field is selected
in the protocol tree window.

The description is the blurb field when available (i.e. not empty),
the name otherwise.

When the item is unselected, the file context (e.g. the file name)
is restored.

Now it is much easier to build a display filter without the manual ;-)


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@2343 f5534014-38df-0310-8fa8-9805f1628bb7