Tvbuffify the HTTP, NNTP, RSH, RTSP, and Telnet dissectors.
Add "tvb_find_line_end()", to find a CR and/or LF-terminated line in a
tvbuff and return its length and the offset of the character after the
line end, for the use of those dissectors.

Add "tvb_strncaseeql()", which is like "tvb_strneql()" except that it
does a case-insensitive comparison.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@2590 f5534014-38df-0310-8fa8-9805f1628bb7