Move "bytes_to_str()" to "strutil.c" from "packet.c" - it's just a string formatter, like "format_text()", and, as "tvbuff.c" now calls it (*vide infra*), we don't want to have to make "tvbuff.c" drag "packet.h" in just to declare "bytes_to_str()".  It's now declared in "strutil.h", so include it in modules that use "bytes_to_str()" and weren't already including it.
Add a "tvb_bytes_to_str()" wrapper that calls "tvb_get_ptr()" to get a
pointer to a chunk of N bytes at a given offset in a tvbuff and then
hands that chunk to "bytes_to_str()".  Convert the code that was doing
that to use "tvb_bytes_to_str()" instead (which caught what I suspect is
a bug in the Q.2931 dissector, where it was handing an offset of 0 to
"tvb_get_ptr()" - a cut-and-pasteo, I think).

Tvbuffify the ARP dissector.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@2634 f5534014-38df-0310-8fa8-9805f1628bb7